Don't parse and format date&time in a loop
The QDateTime code is currently useless, because it converts into the
same format. Implement the same functionality with less and far more
efficient code. Disable rather than remove the old implementation to
preserve timestamp format knowledge for possible future use.

This reduces the time readPacmanLogFile() runs from 5.3 to 2.8 seconds
with Qt 5 and from 3.1 to 2.5 seconds with Qt 4 on my system.